refactor: extract RateDialog into separate class
